Added OOP
App.js - Realized that number of events were not changing when implementing numbers between 1-32, had to change code to make number of events show correctly.
Alert.js - Added info Alert and Error Alert.
CitySearch.js - Implemented InfoAlert to show if city name is typed incorrectly to show alert that city name is not correct.
NumberOfEvents.js - Implemented ErrorAlert to show error if number is not between 1-32 when searching for number of events.